pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "aind_exaspim_register_cells"
description = "Generated from aind-library-template"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Allen Institute for Neural Dynamics"}
]
classifiers = [
    "Programming Language :: Python :: 3"
]
readme = "README.md"
dynamic = ["version"]

dependencies = [
    "numpy==1.23.5",
    "allensdk==2.16.2",
    "zarr==2.18.2",
    "antspyx==0.6.1",
    "scipy==1.10.1",
    "s3fs==2025.5.1",
    "networkx==3.4.2",
    "scikit-learn==1.7.0",
    "pandas==1.5.3"
]

[project.optional-dependencies]
dev = [
    'black',
    'coverage',
    'flake8',
    'interrogate',
    'isort',
    'Sphinx',
    'furo'
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "aind_exaspim_register_cells.__version__"}

[tool.black]
line-length = 79
target_version = ['py310']
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | build
    | dist
    | .conda
  )/
  | .gitignore
)
'''

[tool.coverage.run]
omit = ["*__init__*"]
source = ["aind_exaspim_register_cells", "tests"]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
    "^from .* import .*",
    "^import .*",
    "pragma: no cover"
]
fail_under = 100

[tool.isort]
line_length = 79
profile = "black"
skip = [".conda"]

[tool.interrogate]
exclude = ["setup.py", "docs", "build", ".conda"]
fail-under = 100
